+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - "Inline Function" now works on async function without creating a double `await`. Thanks @automatensalat for catching this!
 - Fixed an issue with "Convert to Template Literal" and backticks. Thanks @fabien0102!
+- Extract Interface doesn't remove existing interface implements anymore. Thanks @BuZZ-dEE!
 
 ### Changed